Add files via upload
Stop Watch / Clock

A stopwatch/clock application combines the functions of a stopwatch and a clock. It allows users to
measure elapsed time with start, stop, and reset features, while also displaying the current time. The
interface typically shows hours, minutes, and seconds, providing a versatile tool for timing activities and
checking the time simultaneously.It also has a timer option that one the deadline is completed its rings an alarm